requirements.txt
FastAPI
uvicorn
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose
email-validator
databases
asyncpg
pytest
httpx
PyJWT
pytest-asyncio
python-dotenv